% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DG_Functions.R
\name{dgConstraint}
\alias{dgConstraint}
\title{Transform MCMC coefficients to satisfy Draper-Gittoes linear constraint}
\usage{
dgConstraint(coef_mat, count_vec)
}
\arguments{
\item{coef_mat}{<IXp> coefficient matrix, with I MCMCM iterations.
The first column should be the overall intercept.}

\item{count_vec}{<px1> vector of integers, total number of observations
per institution or group.}
}
\description{
Transforms coefficients to satisfy the Draper-Gittoes constraint (dot product of the Z-scores and
the count_vec is constrained to be 0). Written by Daniel Kirsner
}
