% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBabyMonitor.R
\name{fitBabyMonitor}
\alias{fitBabyMonitor}
\title{Fit Baby-MONITOR for CPQCC/VON}
\usage{
fitBabyMonitor(minimal_data, num_cat, num_cont, subset = FALSE,
  subset_base_catgory = 1, var_intercept = 40, var_cat = 10,
  var_cat_interaction = 10, var_cont = 10, iters = 100, burn_in = 100,
  alpha = 0.01, n_cutoff = 5, bonferroni = TRUE, t_scores = TRUE,
  outcome_na = "set0", subset_na = "category", cat_na = "category",
  cont_na = "median", score_type = "stat_z_scaled", dat_out = FALSE)
}
\arguments{
\item{minimal_data}{Data_frame with a particular format:

1st column: Outcome vector (0-1 encoding)

2nd column: Institution ID

3rd column: Subset (if subset == TRUE):

Next: num_cat columns of categorical variables (num_cat can equal 0)

Next: num_cont columns of continuous variables (num_cont can equal 0)}

\item{num_cat}{integer. Number of categorical risk adjusters.}

\item{num_cont}{integer. Number of continious risk adjusters.}

\item{subset}{logical; if TRUE perform analysis with a subset variable ( and \code{minimal_data} must have subset data in the 3rd column)}

\item{var_intercept}{scalar. Prior variance for the intercept.}

\item{var_cat}{scalar. Prior variance for categorical risk adjuster parameters.}

\item{var_cat_interaction}{scalar. Prior variance for interactions between categorical risk adjusters.}

\item{var_cont}{scalar. Prior variance for continuous risk adjusters.}

\item{iters}{integer. Desired number of posterior MCMC iterations.}

\item{burn_in}{integer. Number of 'burn-in' MCMC iterations to discard.}

\item{alpha}{scalar in (0,1). Statistical signifiance threshold. Posterior (1-alpha)\% intervals are generated.}

\item{bonferroni}{logical; if TRUE, posterior intervals are widened with the Bonferroni correction.}

\item{t_scores}{logical; if TRUE, posterior intervals confidence intervals constructed with the student t-distribution. If FALSE, z-scores are used.}

\item{outcome_na}{Method for handling any NA values in the outcome vec. 'remove' removes rows with NA outcomes while 'set0' keeps the row and sets the outcome to 0.}

\item{subset_na}{Method for handling any NA subset values. 'remove' removes rows with NA subset values while 'category' makes a new subset category (coded as 99) for NA values.}

\item{cat_na}{Method for handling any NA values in the categorical risk adjusters. 'remove' removes rows with NA values while 'category' makes a new category (coded as 99) for NA catorical risk adjusters.}

\item{cont_na}{Method for handling any NA values in the continous risk adjusters. 'remove' removes rows with NA values while 'median' replaces NA with the median value of the risk adjuster.}

\item{dat_out}{logical; if TRUE, export MCMC iterations and other parameters in the \code{dat} component.}
}
\value{
Returns a large list with the following components:

	inst_mat: Matrix (one row per institution) computing summary statistics, DGH institution ranking, and intervals with both effect-size and standardized scores.

	     full_subset_mat_baseline, full_subset_mat_nobaseline: Matrix with rankings and intervals for the various subset categories.


    subset_baseline_mat, subset_nobaseline_mat: Matrix with rankings and intervals for subset categories within institution.


	group_labels: A vector of the names of each institution

    mcmc_fit: Matrix of MCMC iterations for each coefficient
	dg_z: Matrix of computed z score for each institution at each MCMC iterations.

	coefs: Names of each coefficient (1st is intercept, then institution, then everything else)

	prior_var_vec: Vector of prior variances for each coefficient

	model_matrix: Design matrix
}
\description{
\code{fitBabyMonitor} applies the Baby-MONITOR score  to a single performance indicator. Designed for CPQCC/VON usage.
}
