% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMatrix.R
\name{modelMatrix}
\alias{modelMatrix}
\title{Create a sparse design matrix from data frame}
\usage{
modelMatrix(x, interactions = FALSE, intercept = FALSE)
}
\arguments{
\item{x}{A data.frame with finite, non-missing elements.}

\item{interactions}{Logical; if \code{TRUE} then all interactions included.}

\item{intercept}{Logcal; if \code{TRUE} then include a leading column of 1's.}
}
\description{
\code{modelMatrix} takes in a data.frame and encodes it as a sparse model.Matrix object.
Factors use treatment ("one-hot") encoding, creating indicator variables from categorical
and binary variables.
Note that names of categorical variables will change slightly 
(n-1 unique values appended to name)
}
