% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreComposite.r
\name{scoreComposite}
\alias{scoreComposite}
\title{Compute composite Baby-MONITOR scores}
\usage{
scoreComposite(returner_list, alpha = 0.05, bonferroni = TRUE)
}
\arguments{
\item{alpha}{scalar in (0,1). Statistical signifiance threshold. Posterior (1-alpha)\% intervals are generated.}

\item{bonferroni}{logical; if TRUE, posterior intervals are widened with the Bonferroni correction.}

\item{returner_list:}{List of outputs from fitBabyMonitor(), one element for each performance indicator.}
}
\description{
\code{scoreComposite} generates composite Baby-MONITOR scores  after analysis using fitBabyMonitor() on multiple performance indicators.
}
