% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBabyMonitor.R
\name{fitBabyMonitor}
\alias{fitBabyMonitor}
\title{Fit Baby-MONITOR for CPQCC/VON}
\usage{
fitBabyMonitor(minimal_data, num_cat, num_cont, subset = FALSE,
  var_intercept = 40, var_cat = 10, var_cat_interaction = 10,
  var_cont = 10, iters = 100, burn_in = 100, alpha = 0.05,
  verbose = TRUE, outcome.na = "set0", subset.na = "category",
  cat.na = "category", cont.na = "median")
}
\arguments{
\item{minimal_data}{Data_frame with a particular format:

1st column: Outcome vector (0-1 encoding)

2nd column: Institution ID

3rd column: Subset (if subset == TRUE):

Next: num_cat columns of categorical variables (num_cat can equal 0)

Next: num_cont columns of continuous variables (num_cont can equal 0)}

\item{num_cat}{Scalar number of categorical variables.}

\item{num_cont}{Number of continuous variables.}

\item{subset}{If TRUE, perform subset fitting tasks}

\item{var_intercept}{Prior variance for intercept parameter.}

\item{var_cat}{Prior variance for categorical parameters.}

\item{var_cat_interaction}{Prior variance for categorical interaction parameters.}

\item{var_cont}{Prior variance for continuous parameters.}

\item{iters}{Number of MCMC iterations to use.}

\item{burn_in}{Number of initial iterations to discard for burn in.}

\item{alpha}{We look at posterior (1-alpha)\% posterior intervals.}

\item{verbose}{If TRUE, display status messages while fitting}

\item{sparse}{Should design_matrix be stored as Sparse matrix? Requires Matrix package.}
}
\value{
Returns a large list with the following components:

	inst_mat: Matrix (one row per institution) computing summary statistics, DGH institution ranking, and intervals with both effect-size and standardized scores.

	     full_subset_mat_baseline, full_subset_mat_nobaseline: Matrix with rankings and intervals for the various subset categories.


    subset_baseline_mat, subset_nobaseline_mat: Matrix with rankings and intervals for subset categories within institution.


	group_labels: A vector of the names of each institution

    mcmc_fit: Matrix of MCMC iterations for each coefficient
	dg_z: Matrix of computed z score for each institution at each MCMC iterations.

	coefs: Names of each coefficient (1st is intercept, then institution, then everything else)

	prior_var_vec: Vector of prior variances for each coefficient

	model_matrix: Design matrix
}
\description{
\code{fitBabyMonitor} comprehensively applies the Baby-MONITOR score
Returns a large list. Use inst_mat for institution rankings,
full_subset_mat_baseline and full_subset_mat_nobaseline for subset rankings,
and subset_baseline_mat and subset_nobaseline_mat for subset rankings within institution.
}
